requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
numpy>=1.24.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
Pillow>=10.0.0
EOF